Full API

Autogenerated API list

CavityWaveguide.CavityWaveguideAbsorption — Type

CavityWaveguideAbsorption{B1,B2} <: CavityWaveguideOperator{B1,B2}

Structure for fast simultaneous cavity creation and waveguide annihilation operator

CavityWaveguide.CavityWaveguideEmission — Type

CavityWaveguideEmission{B1,B2} <: CavityWaveguideOperator{B1,B2}

Structure for fast simultaneous cavity annihilation and waveguide creation operator

CavityWaveguide.CavityWaveguideOperator — Type

Abstract type used for operators on acting on a combined WaveguideBasis and cavity basis (FockBasis)

CavityWaveguide.TwophotonTimestepView — Type

TwophotonTimestepView{T}

Structure for viewing slice along same times in twophoton state.

CavityWaveguide.TwophotonView — Type

TwophotonView{T} <: AbstractMatrix{T}

Structure for viewing twophoton state as symmetric matrix (only upper triangluar part is stored in memory). Returned from view_twophoton.

CavityWaveguide.WaveguideBasis — Type

WaveguideBasis(N, times)

Basis for time binned Waveguide where N is the number of photons in the waveguide. Currently restricted to either 1 or 2. Times is timeinterval over which the photon state should be binned.

CavityWaveguide.WaveguideCreate — Type

WaveguideCreate{N} <: WaveguideOperator

Operator structure for dispatching creation operation on Waveguide state. N is used to dispatch one or two photon routine.

CavityWaveguide.WaveguideDestroy — Type

WaveguideDestroy{N} <: WaveguideOperator

Operator structure for dispatching annihilation operation on Waveguide state. N is used to dispatch one or two photon routine.

CavityWaveguide.WaveguideOperator — Type

Abstract class for WaveguideOperators. Used to dispatch special mul! function.

CavityWaveguide.absorption — Method

absorption(b1::WaveguideBasis{T},b2::FockBasis) where T
absorption(b1::FockBasis,b2::WaveguideBasis{T}) where T

Create CavityWaveguideAbsorption that applies create(b::FockBasis) on FockBasis and destroy(b::WaveguideBasis{T}) on WaveguideBasis{T}.

CavityWaveguide.emission — Method

emission(b1::WaveguideBasis{T},b2::FockBasis) where T
emission(b1::FockBasis,b2::WaveguideBasis{T}) where T

Create CavityWaveguideEmission that applies destroy(b::FockBasis) on FockBasis and create(b::WaveguideBasis{T}) on WaveguideBasis{T}.

CavityWaveguide.get_nsteps — Method

get_nsteps(basis::WaveguideBasis)
get_nsteps(basis::Basis)
get_nsteps(basis::CompositeBasis)

Return nsteps of WaveguideBasis given either a WaveguideBasis or a CompositeBasis containing a WaveguideBasis

CavityWaveguide.get_waveguide_basis — Method

get_waveguide_basis(basis::CompositeBasis)

Returns WaveguideBasis from CompositeBasis.bases

CavityWaveguide.get_waveguide_location — Method

get_waveguide_location(basis::WaveguideBasis)
get_waveguide_location(basis::CompositeBasis)

Return index of WaveguideBasis location in Hilbert space of basis b. Btotal = BW ⊗ BC where BW is a WaveguideBasis and BC some other basis then get_waveguide_location(Btotal) returns 1. While Btotal = BC ⊗ BW with get_waveguide_location(Btotal) returns 2.

CavityWaveguide.get_waveguidetimeindex — Method

get_waveguidetimeindex(basis::WaveguideBasis)
get_waveguidetimeindex(basis::Basis)
get_waveguidetimeindex(basis::CompositeBasis)

Return timeindex of WaveguideBasis given either a WaveguideBasis or a CompositeBasis containing a WaveguideBasis

CavityWaveguide.onephoton — Method

onephoton(b::WaveguideBasis,ξ::Function,times,args...,norm=True)
onephoton(b::WaveguideBasis,ξvec;norm=true)

Create a onephoton wavepacket of the form $W^†(ξ) |0⟩ = \int_{t_0}^{t_{end}} dt ξ(t) w^†(t) |0⟩$. Here $t_0=0$ and $t_{end}$ is determined by WaveguideBasis. ξ is a function evaluated as ξ.(times,args...). ξvec is a vector of length: b.nsteps. If norm==true the state is normalized through normalize!.

CavityWaveguide.twophoton — Method

twophoton(b::WaveguideBasis,ξ::Function,times,args...,norm=True)
twophoton(b::WaveguideBasis,ξvec::Matrix;norm=true)

Create a twophoton wavepacket of the form $\int_{t_0}^{t_{end}} dt' \int_{t_0}^{t_{end}} dt ξ(t,t') w^†(t)w^†(t') |0⟩$. Here $t_0=0$ and $t_{end}$ is determined by WaveguideBasis. ξ is a function evaluated as ξ(t1,t2,args...). ξvec is a matrix of dimension: (b.nsteps,b.nsteps), where ξvec[i,j] = ξ(times[i],times[j]), where times is defined in WaveguideBasis.

CavityWaveguide.view_onephoton — Method

view_onephoton(ψ::Ket)
view_onephoton(ψ::Ket,index)

Return a view of the onephoton mode $ξ(t)$ given an input state containing a onephoton waveguide state: $\int_{t_0}^{t_{end}} dt ξ(t) w^†(t) |0⟩$ If no index is provided the ground state is returned. Index should follow same form outlined in view_waveguide.

TO DO: PERHAPS CHANGE NAME?

CavityWaveguide.view_twophoton — Method

view_twophoton(ψ::Ket)
view_twophoton(ψ::Ket,index)

Return a view of the twophoton mode $ξ(t_1,t_2)$ given an input state containing a twophoton waveguide state: $\int_{t_0}^{t_{end}} dt' \int_{t_0}^{t_{end}} dt ξ(t,t') w^†(t)w^†(t') |0⟩$ If no index is provided the ground state is returned. Index should follow same form outlined in view_waveguide.

TO DO: PERHAPS CHANGE NAME?

CavityWaveguide.view_waveguide — Method

view_waveguide(ψ::ket)
view_waveguide(ψ::ket,index)

View the Waveguide state given a state ψ containing a WaveguideBasis by returning view(reshape(ψ.data,Tuple(ψ.basis.shape)),index...). If no index is provided the ground state is returned. The index provided should be of the form [:,i,j] where (:) is at the location of the WaveguideBasis and i and j are indeces of other basises. See example:

times=0:0.1:10
bw = WaveguideBasis(2,times)
bc1 = FockBasis(2)
bc2 = FockBasis(2)
ψ_waveguide = onephoton(bw,x->1)
ψ_total = ψ_waveguide ⊗ fockstate(bc1,1) ⊗ fockstate(bc2,1)
ψ_view = view_waveguide(ψ_total)
ψ_view_index = view_waveguide(ψ_total,[:,1,1])
ψ_view==ψ_view_index

CavityWaveguide.waveguide_evolution — Method

waveguide_evolution(tspan, psi0, H; fout)

Integrate time-dependent Schroedinger equation to evolve states or compute propagators.

Arguments

tspan: Vector specifying the points of time for which output should be displayed.

psi0: Initial state vector can only be a ket.

H: Operator containing a WaveguideOperator either through a LazySum or LazyTensor.

fout=nothing: If given, this function fout(t, psi) is called every time step. Example: fout(t,psi) = expect(A,psi) will return the epectation value of A at everytimestep. ATTENTION: The state psi is neither normalized nor permanent! It is still in use by the ode solver and therefore must not be changed.

Output

if fout=nothing the output of the solver will be the state ψ at the last timestep.

if fout is given a tuple with the state ψ at the last timestep and the output of fout is given. If fout returns a tuple the tuple will be flattened.

Example fout(t,psi) = (expect(A,psi),expect(B,psi)) will result in a tuple (ψ, ⟨A(t)⟩,⟨B(t)⟩), where ⟨A(t)⟩ is a vector with the expectation value of A as a function of time.

CavityWaveguide.waveguide_montecarlo — Method

waveguide_montecarlo(times,psi,H,J;fout=nothing)

CavityWaveguide.zerophoton — Method

zerophoton(bw::WaveguideBasis)

Create a waveguide vacuum state |0⟩

LinearAlgebra.mul! — Function

mul!(result::Ket{B1}, a::LazyTensor{B1,B2,F,I,T}, b::Ket{B2}, alpha, beta)
mul!(result::Bra{B1}, a::Bra{B2}, b::LazyTensor{B1,B2,F,I,T}, alpha, beta)

In-place multiplication of operators/state vectors. Updates result as result = alpha*a*b + beta*result. a is a LazyTensor that contains a WaveguideOperator

mul!(result::Ket{B1}, a::CavityWaveguideOperator, b::Ket{B2}, alpha, beta) where {B1<:Basis,B2<:Basis}
mul!(result::Bra{B1}, a::Bra{B2}, b::CavityWaveguideOperator, alpha, beta) where {B1<:Basis,B2<:Basis}

Fast in-place multiplication of operators/state vectors. Updates result as result = alpha*a*b + beta*result. a is a CavityWaveguideOperator. Routine only works if [WaveguideBasis] is the first or last basis in Hilbertspace.

QuantumOpticsBase.create — Function

create(basis::WaveguideBasis{1})
create(basis::WaveguideBasis{2})

Creation operator for WaveguideBasis for either one or two photons.

QuantumOpticsBase.destroy — Function

destroy(basis::WaveguideBasis{1})
destroy(basis::WaveguideBasis{2})

Annihilation operator for WaveguideBasis for either one or two photons.

QuantumInterface.dagger — Function

dagger(basis::WaveguideBasis{1})
dagger(basis::WaveguideBasis{2})

Dagger opration on Waveguide operator.

QuantumInterface.tensor — Function

tensor(op1::AbstractOperator,op2::CavityWaveguideAbsorption)
tensor(op1::CavityWaveguideAbsorption,op2::AbstractOperator)
tensor(op1::AbstractOperator,op2::CavityWaveguideEmission)
tensor(op1::CavityWaveguideEmission,op2::AbstractOperator)

Methods for tensorproducts between QuantumOptics.jl operators and CavityWaveguideOperator.

QuantumOpticsBase.identityoperator — Function

identityoperator(a::WaveguideOperator)

Return identityoperator(a.basisl). QUESTION: (does basisl or basis_r matter?)

identityoperator(a::CavityWaveguideOperator)

Return identityoperator(a.basisl). QUESTION: (does basisl or basis_r matter?)