Full API

Autogenerated API list

WaveguideQED.CavityWaveguideAbsorption — Type

CavityWaveguideAbsorption{B1,B2} <: CavityWaveguideOperator{B1,B2}

Structure for fast simultaneous Cavity creation and Waveguide annihilation operator

WaveguideQED.CavityWaveguideEmission — Type

CavityWaveguideEmission{B1,B2} <: CavityWaveguideOperator{B1,B2}

Structure for fast simultaneous cavity annihilation and waveguide creation operator

WaveguideQED.CavityWaveguideOperator — Type

Abstract type used for operators on acting on a combined WaveguideBasis and cavity basis (FockBasis)

WaveguideQED.Detector — Type

Detector(wa,wb)

Detector operation defined by giving waveguide annihilation operator wa and wb from two subsystems. wa acts on the first subsystem of a LazyTensorKet or LazySumKet and wb on the second.

WaveguideQED.LazySumKet — Type

LazySumKet(kets...)

Lazy sum of LazyTensorKets that is used to express entanglement between subsystems in LazyTensorKets.

WaveguideQED.LazyTensorBra — Type

LazyTensorBra(bras)

Lazy tensor product between bras. Used in functions for beamsplitter and subsequent detection.

WaveguideQED.LazyTensorKet — Type

LazyTensorKet(kets)

Lazy tensor product between kets. Used in functions for beamsplitter and subsequent detection.

WaveguideQED.OnePhotonView — Method

OnePhotonView(ψ::T) where {T<:SingleWaveguideKet}
OnePhotonView(ψ::T,index::I) where {T<:SingleWaveguideKet,I<:Union{Vector{Any},Vector{Int64},Tuple{Vararg{Union{Int64,Colon}}}}}
OnePhotonView(ψ::T,WI::Int)  where {T<:MultipleWaveguideKet}
OnePhotonView(ψ::T,index::I,WI::Int) where {T<:MultipleWaveguideKet,I<:Union{Vector{Any},Vector{Int64},Tuple{Vararg{Union{Int64,Colon}}}}}

Return a view of the onephoton wavefunction $ξ(t)$ given a state containing a WaveguideBasis. If the WaveguideBasis contains more than one waveguide, a waveguide index $WI$ is required to indicate which waveguide is viewed (1,2,3,... etc.) If the state contains more basises (e.g. a cavity) $index$ used to indicate which state should be viewed. Index should follow same form outlined in view_waveguide and if not given the groundstate is assummed.

WaveguideQED.TwoPhotonTimestepView — Type

TwoPhotonTimestepView{T}

Structure for viewing slice along same times in twophoton state.

WaveguideQED.TwoPhotonView — Type

TwophotonView{T} <: AbstractMatrix{T}

Structure for viewing twophoton state as symmetric matrix (only upper triangluar part is stored in memory).

WaveguideQED.TwoPhotonView — Method

TwoPhotonView(ψ::T) where {T <: SingleWaveguideKet}
TwoPhotonView(ψ::T,index::I) where {T <: SingleWaveguideKet,I<:Union{Vector{Any},Vector{Int64},Tuple{Vararg{Union{Int64,Colon}}}}}
TwoPhotonView(ψ::T,WI::Int) where {T <: MultipleWaveguideKet}
TwoPhotonView(ψ::T,WI1::Int,WI2::Int) where {T <: MultipleWaveguideKet}
TwoPhotonView(ψ::T,WI::I) where {T <: MultipleWaveguideKet,I<:Union{Vector{Int64},Tuple{Vararg{Int64}}}}
TwoPhotonView(ψ::T,index::I,WI::Int)  where {T<:MultipleWaveguideKet,I<:Union{Vector{Any},Vector{Int64},Tuple{Vararg{Union{Int64,Colon}}}}}

Return a view of the twophoton wavefunction $ξ(t1,t2)$ given a state containing a WaveguideBasis. If the WaveguideBasis contains more than one waveguide, a waveguide index $WI$ is required to indicate which waveguide is viewed. WI follows the same syntax outlined in twophoton for more information on how to view the state. If the state contains more basises (e.g. a cavity) $index$ used to indicate which state should be viewed. Index should follow same form outlined in view_waveguide and if not given the groundstate is assummed.

WaveguideQED.TwoWaveguideTimestepView — Type

TwoWaveguideTimestepView{T} <:AbstractVector{T}

Structure for viewing slice along same times in one photon Left Right state.

WaveguideQED.TwoWaveguideView — Type

TwoWaveguideView{T} <: AbstractMatrix{T}

Structure for viewing state with one photon in Left and Right waveguide.

WaveguideQED.WaveguideBasis — Type

WaveguideBasis(Np,Nw, times)
aveguideBasis(Np, times)

Basis for time binned Waveguide where Np is the number of photons in the waveguide and Nw the number of waveguides (default is 1). . Currently number of photons is restricted to either 1 or 2. Times is timeinterval over which the photon state should be binned.

WaveguideQED.WaveguideCreate — Type

WaveguideCreate{B1,B2,N,idx} <: WaveguideOperator{B1,B2}

Operator structure for dispatching creation operation on Waveguide state. Np is used to dispatch one or two photon routine and idx denotes the index of the waveguide the operator is acting on.

WaveguideQED.WaveguideDestroy — Type

WaveguideDestroy{B1,B2,Np,idx} <: WaveguideOperator{B1,B2}

Operator structure for dispatching annihilation operation on Waveguide state. Np is used to dispatch one or two photon routine and idx denotes the index of the waveguide the operator is acting on.

WaveguideQED.WaveguideOperator — Type

Abstract class for WaveguideOperators. Used to dispatch special mul! function.

QuantumInterface.dagger — Method

dagger(op::WaveguideCreate)
dagger(op::WaveguideCreate)

Dagger opration on Waveguide operator.

QuantumInterface.tensor — Method

tensor(a::AbstractOperator,b::CavityWaveguideAbsorption)
tensor(a::CavityWaveguideAbsorption,b::AbstractOperator)
tensor(a::AbstractOperator,b::CavityWaveguideEmission)
tensor(a::CavityWaveguideEmission,b::AbstractOperator)

Methods for tensorproducts between QuantumOptics.jl operators and CavityWaveguideOperator.

QuantumOpticsBase.create — Method

create(basis::WaveguideBasis{Np,1}) where {Np}
create(basis::WaveguideBasis{Np,Nw},i::Int) where {Np,Nw}

Creation operator $w^\dagger$ for WaveguideBasis $w_i(t_k)^\dagger | \emptyset \rangle = | 1_k \emptyset \rangle_i$ with cutoff (maximum number of photons Np) where i is the index of the waveguide. $t_k$ is determined by the timeindex property of the operator which can be changed by set_waveguidetimeindex!(a::WaveguideOperator,k::Int)

Arguments

basis of type WaveguideBasis, defines cutoff photon number Np and number of waveguides Nw

i determines which waveguide the operator acts on and should be i ≤ Nw. If Nw=1 then i=1 is assumed (there is only one waveguide).

Returns

WaveguideCreate{B,B,Np,i} <: WaveguideOperator

QuantumOpticsBase.destroy — Method

destroy(basis::WaveguideBasis{Np,1}) where {Np}
destroy(basis::WaveguideBasis{Np,Nw},i::Int) where {Np,Nw}

Annihilation operator $w$ for WaveguideBasis $w_i(t_k) | 1_j \emptyset \rangle_i = \delta_{k,j} | \emptyset \rangle$ with cutoff (maximum number of photons Np) where i is the index of the waveguide. $t_k$ is determined by the timeindex property of the operator which can be changed by set_waveguidetimeindex!(a::WaveguideOperator,k::Int)

Arguments

basis of type WaveguideBasis, defines cutoff photon number Np and number of waveguides Nw

i determines which waveguide the operator acts on and should be i ≤ Nw. If Nw=1 then i=1 is assumed (there is only one waveguide).

Returns

WaveguideDesroy{B,B,Np,i} <: WaveguideOperator

QuantumOpticsBase.identityoperator — Method

identityoperator(a::CavityWaveguideOperator)

Return identityoperator(a.basis_l).

QuantumOpticsBase.identityoperator — Method

identityoperator(a::WaveguideOperator)

Return identityoperator(a.basis_l).

WaveguideQED.absorption — Method

absorption(b1::WaveguideBasis{T},b2::FockBasis) where T
absorption(b1::FockBasis,b2::WaveguideBasis{T}) where T

Create CavityWaveguideAbsorption that applies create(b::FockBasis) on FockBasis and destroy(b::WaveguideBasis{T}) on WaveguideBasis{T}.

WaveguideQED.detect_double_click — Method

detect_double_click(ψ,detector1,detector2,projection)
detect_double_click(ψ,detector1,detector2)

Calculate probability of observing projection after beamsplitter operation and two subsequent detection events defined in detector1 and detector2 on the state ψ.

Arguments

ψ can be either LazyTensorKet or LazySumKet and is the state on which the beamsplitter and detection is applied

detector1 defines the first beamsplitter and subsequent detection operation. See Detector for more details on how to define.

detector2 defines the second beamsplitter and subsequent detection operation. See Detector for more details on how to define.

projection if given is a LazyTensorKet or LazySumKet which projects onto the state after the measurement. If no projection is given, instead the total probability of having the detector click is given by applying all possible combinations of projections using get_all_projectors.

Returns

If projection is given: Returns probability of having detector1 and detector2 click and being in state defined by projection

If projection is not given: Returns the total probability of having detector1 and detector2 click by applying all possibile projections with zerophotons in the waveguide using get_all_projectors.

See Beamsplitter for an example on how to use.

WaveguideQED.detect_single_click — Method

detect_single_click(ψ,detector::Detector,projection)
detect_single_click(ψ,detector::Detector)

Calculate probability of observing projection after beamsplitter operation and subsequent detection event defined in detector on the state ψ.

Arguments

ψ can be either LazyTensorKet or LazySumKet and is the state on which the beamsplitter and detection is applied

detector defines the beamsplitter and subsequent detection operation. See Detector for more details on how to define.

projection if given is a LazyTensorKet or LazySumKet which projects onto the state after the measurement. If no projection is given, instead the total probability of having the detector click is given by applying all possible combinations of projections using get_all_projectors.

Returns

If projection is given returns probability of having detector click and being in state defined by projection

If projection is not given returns the total probability of having a the detector click (only a single click, for double clicks use detect_double_click) by applying all possibile projections with zerophotons in the waveguide using get_all_projectors.

See Beamsplitter for an example on how to use.

WaveguideQED.emission — Method

emission(b1::WaveguideBasis{T},b2::FockBasis) where T
emission(b1::FockBasis,b2::WaveguideBasis{T}) where T

Create CavityWaveguideEmission that applies destroy(b::FockBasis) on FockBasis and create(b::WaveguideBasis{T}) on WaveguideBasis{T}.

WaveguideQED.get_all_projectors — Method

get_all_projectors(b)

Returns all combinations of possible states with zerophotons in the waveguide.

WaveguideQED.get_nsteps — Method

get_nsteps(basis::WaveguideBasis)
get_nsteps(basis::Basis)
get_nsteps(basis::CompositeBasis)

Return nsteps of WaveguideBasis given either a WaveguideBasis or a CompositeBasis containing a WaveguideBasis

WaveguideQED.get_number_of_waveguides — Method

get_number_of_waveguides(basis::WaveguideBasis)
get_number_of_waveguides(basis::Basis)
get_number_of_waveguides(basis::CompositeBasis)

Return number of waveguides Nw of WaveguideBasis{Np,Nw} given either a WaveguideBasis{Np,Nw} or a CompositeBasis containing a WaveguideBasis{Np,Nw}

WaveguideQED.get_waveguide_basis — Method

get_waveguide_basis(basis::CompositeBasis)

Returns WaveguideBasis from CompositeBasis.bases

WaveguideQED.get_waveguide_location — Method

get_waveguide_location(basis::WaveguideBasis)
get_waveguide_location(basis::CompositeBasis)

Return index of WaveguideBasis location in Hilbert space of basis b. Btotal = waveguidebasis ⊗ otherbasis where waveguidebasis is a WaveguideBasis and otherbasis some other basis then get_waveguide_location(Btotal) returns 1. While Btotal = otherbasis ⊗ waveguidebasis with get_waveguide_location(Btotal) returns 2.

WaveguideQED.get_waveguide_operators — Method

get_waveguide_operators(basis::LazySum)
get_waveguide_operators(basis::LazyProduct)
get_waveguide_operators(basis::LazyTensor)
get_waveguide_operators(basis::Tuple)
get_waveguide_operators(basis::Array)
get_waveguide_operators(basis::WaveguideOperator)

Returns all WaveguideOperator in LazyOperator or from a list of operators. If no WaveguideOperator is found, and empty array is returned.

WaveguideQED.get_waveguidetimeindex — Method

get_waveguidetimeindex(op)

Return timeindex of operator or list of operators containing WaveguideOperator and assert that all timeindeces are the same.

WaveguideQED.onephoton — Method

onephoton(b::WaveguideBasis{T,1},ξ::Function,times,args...,norm=True) where {T}
onephoton(b::WaveguideBasis{T,1},ξvec;norm=true) where {T}
onephoton(b::WaveguideBasis{T,Nw},i::Int,ξ::Function,times,args...; norm=true) where {T,Nw}
onephoton(b::WaveguideBasis{T,Nw},i,ξvec;norm=true) where {T,Nw}

Create a onephoton wavepacket in waveguide of the form $W^\dagger(\xi) |0 \rangle = \int_{t_0}^{t_{end}} dt \xi(t) w_{\mathrm{i}}^\dagger(t) |\emptyset \rangle$.

Arguments

b::WaveguideBasis{T,Nw}: the basis of the waveguides, where T is the number of photons in the waveguides and Nw is the number of waveguides.

i::Int (optional): the index of the waveguide where the photon is created. If not provided, i=1 is assumed.

ξ::Function or ξ::AbstractArray: the wavefunction of the state. Can be a function with structure ξ.(times,args...) or an AbstractArray of length b.nsteps.

times: A vector or range of times where the wavefunction is evaluated, used only if ξ is a function.

args...: additional arguments to be passed to ξ if it is a function.

norm::Bool=true: normalize the resulting wavepacket.

Returns

[Ket(b)]: a ket with the wavefunction defined above.

If b only contains one waveguide, the output wavefunction will contain on excitation in the Waveguide (i=1). If b contains multiple waveguides and only one index is given (i or j), then i=j is assumed.

WaveguideQED.plot_twophoton! — Method

plot_twophoton!(ax,twophotonstate::TwophotonView,times)
plot_twophoton!(ax,twophotonstate::TwoWaveguideView,times)
plot_twophoton!(ax,state::Ket,times)
plot_twophoton!(ax,twophotonstate,times)

Plots the twophoton state in the given ax.

Arguments

ax of type PyObject <AxesSubplot: > from PyPlot

State to be plotted twophotonstate or state. If state is a Ket TwoPhotonView is called to extract twophotonstate. Otherwise twophotonstate should be AbstractArray of dimensions (length(times),length(times)).

#Return ax.contour object with the plotted state.

WaveguideQED.set_waveguidetimeindex! — Method

set_waveguidetimeindex!(op,index)

Set timeindex of all WaveguideOperator in operator or list of operators to index

WaveguideQED.twophoton — Method

twophoton(b::WaveguideBasis{T,1},ξ::Function,times,args...,norm=True) where {T}
twophoton(b::WaveguideBasis{T,1},ξvec::Matrix;norm=true) where {T}
twophoton(b::WaveguideBasis{T,Nw},i::Int,ξ::Function,times,args...;norm=true) where {T,Nw}
twophoton(b::WaveguideBasis{T,Nw},i::Int,ξ::Matrix;norm=true) where {T,Nw}
twophoton(b::WaveguideBasis{T,Nw},i::Int,j::Int,ξ::Function,times,args...;norm=true) where {T,Nw}
twophoton(b::WaveguideBasis{T,Nw},i::Int,j::Int,ξ::Matrix;norm=true) where {T,Nw}
twophoton(b::WaveguideBasis{T,Nw},WI::I,ξ::Matrix;norm=true) where {T,Nw,I<:Union{Vector{Int64},Tuple{Vararg{Int64}}}}
twophoton(b::WaveguideBasis{T,Nw},WI::I,ξ::Function,times,args...;norm=true)

Create a twophoton wavepacket of the form $\int_{t_0}^{t_{end}} dt' \int_{t_0}^{t_{end}} dt \xi(t,t') w_{\mathrm{i}}^\dagger(t)w_{\mathrm{j}}^\dagger(t') |\emptyset \rangle$.

Arguments

b::WaveguideBasis{T,Nw}: the basis of the waveguides, where T is the number of photons in the waveguides and Nw is the number of waveguides.

i::Int (optional): the index of the waveguide where the first photon is created. If not provided, i=1 is assumed.

j::Int (optional): the index of the waveguide where the second photon is created. If not provided, j=i is assumed.

ξ::Function or ξ::Matrix: the wavefunction of the state. Can be a function with structure ξ(times[l],times[m],args...) or a matrix of dimension (b.nsteps, b.nsteps).

times: A vector or range of times where the wavefunction is evaluated, used only if ξ is a function.

args...: additional arguments to be passed to ξ if it is a function.

norm::Bool=true: normalize the resulting wavepacket.

Returns

[Ket(b)]: a ket with the wavefunction defined above.

If b only contains one waveguide, the output wavefunction will contain two excitations in the same waveguide (i=j=1). If b contains multiple waveguides and only one index is given (i or j), then i=j is assumed, and two excitations in the same waveguide are returned. i and j can also be given as a tuple or vector WI= (i,j) or WI= [i,j]

WaveguideQED.view_waveguide — Method

view_waveguide(ψ::ket)
view_waveguide(ψ::ket,index)

View the Waveguide state given a state ψ containing a WaveguideBasis by returning view(reshape(ψ.data,Tuple(ψ.basis.shape)),index...). If no index is provided the ground state is returned. The index provided should be of the form [:,i,j] where (:) is at the location of the WaveguideBasis and i and j are indeces of other basises. See example:

julia> using WaveguideQED; #hide
julia> using QuantumOptics; #hide
julia> times=0:0.1:10;
julia> bw = WaveguideBasis(2,times);
julia> bc1 = FockBasis(2);
julia> bc2 = FockBasis(2);
julia> ψ_waveguide = onephoton(bw,x->1);
julia> ψ_total = ψ_waveguide ⊗ fockstate(bc1,1) ⊗ fockstate(bc2,1);
julia> ψ_view = view_waveguide(ψ_total);
julia> ψ_view_index = view_waveguide(ψ_total,[:,1,1]);
julia> ψ_view==ψ_view_index

WaveguideQED.waveguide_evolution — Method

waveguide_evolution(tspan, psi0, H; fout)

Integrate time-dependent Schroedinger equation to evolve states or compute propagators.

Arguments

tspan: Vector specifying the points of time for which output should be displayed.

psi0: Initial state vector can only be a ket.

H: Operator containing a WaveguideOperator either through a LazySum or LazyTensor.

fout=nothing: If given, this function fout(t, psi) is called every time step. Example: fout(t,psi) = expect(A,psi) will return the epectation value of A at everytimestep. If fout =1 the state psi is returned for all timesteps in a vector. ATTENTION: The state psi is neither normalized nor permanent! It is still in use by the ode solver and therefore must not be changed.

Returns

if fout=nothing the output of the solver will be the state ψ at the last timestep.

if fout is given a tuple with the state ψ at the last timestep and the output of fout is given. If fout returns a tuple the tuple will be flattened.

if fout = 1 ψ at all timesteps is returned.

Example fout(t,psi) = (expect(A,psi),expect(B,psi)) will result in a tuple (ψ, ⟨A(t)⟩,⟨B(t)⟩), where ⟨A(t)⟩ is a vector with the expectation value of A as a function of time.

WaveguideQED.waveguide_montecarlo — Method

waveguide_montecarlo(times,psi,H,J;fout=nothing)

See documentation for waveguide_evolution. J should be a list of collapse operators.

WaveguideQED.zerophoton — Method

zerophoton(bw::WaveguideBasis)

Create a waveguide vacuum state |0⟩

LinearAlgebra.mul! — Function

mul!(result::Ket{B1}, a::LazyTensor{B1,B2,F,I,T}, b::Ket{B2}, alpha, beta)

In-place multiplication of operators/state vectors. Updates result as result = alpha*a*b + beta*result. a is a LazyTensor that contains a WaveguideOperator

mul!(result::Ket{B1}, a::CavityWaveguideEmission, b::Ket{B2}, alpha, beta) where {B1<:Basis,B2<:Basis}
mul!(result::Ket{B1}, a::CavityWaveguideAbsorption, b::Ket{B2}, alpha, beta) where {B1<:Basis,B2<:Basis}

Fast in-place multiplication of operators/state vectors. Updates result as result = alpha*a*b + beta*result. a is a CavityWaveguideOperator.

QuantumOpticsBase.create — Function

create(basis::WaveguideBasis{Np,1}) where {Np}
create(basis::WaveguideBasis{Np,Nw},i::Int) where {Np,Nw}

Creation operator $w^\dagger$ for WaveguideBasis $w_i(t_k)^\dagger | \emptyset \rangle = | 1_k \emptyset \rangle_i$ with cutoff (maximum number of photons Np) where i is the index of the waveguide. $t_k$ is determined by the timeindex property of the operator which can be changed by set_waveguidetimeindex!(a::WaveguideOperator,k::Int)

Arguments

basis of type WaveguideBasis, defines cutoff photon number Np and number of waveguides Nw

i determines which waveguide the operator acts on and should be i ≤ Nw. If Nw=1 then i=1 is assumed (there is only one waveguide).

Returns

WaveguideCreate{B,B,Np,i} <: WaveguideOperator

QuantumOpticsBase.destroy — Function

destroy(basis::WaveguideBasis{Np,1}) where {Np}
destroy(basis::WaveguideBasis{Np,Nw},i::Int) where {Np,Nw}

Annihilation operator $w$ for WaveguideBasis $w_i(t_k) | 1_j \emptyset \rangle_i = \delta_{k,j} | \emptyset \rangle$ with cutoff (maximum number of photons Np) where i is the index of the waveguide. $t_k$ is determined by the timeindex property of the operator which can be changed by set_waveguidetimeindex!(a::WaveguideOperator,k::Int)

Arguments

basis of type WaveguideBasis, defines cutoff photon number Np and number of waveguides Nw

i determines which waveguide the operator acts on and should be i ≤ Nw. If Nw=1 then i=1 is assumed (there is only one waveguide).

Returns

WaveguideDesroy{B,B,Np,i} <: WaveguideOperator

QuantumInterface.dagger — Function

dagger(op::WaveguideCreate)
dagger(op::WaveguideCreate)

Dagger opration on Waveguide operator.

QuantumInterface.tensor — Function

tensor(a::AbstractOperator,b::CavityWaveguideAbsorption)
tensor(a::CavityWaveguideAbsorption,b::AbstractOperator)
tensor(a::AbstractOperator,b::CavityWaveguideEmission)
tensor(a::CavityWaveguideEmission,b::AbstractOperator)

Methods for tensorproducts between QuantumOptics.jl operators and CavityWaveguideOperator.

QuantumOpticsBase.identityoperator — Function

identityoperator(a::WaveguideOperator)

Return identityoperator(a.basis_l).

identityoperator(a::CavityWaveguideOperator)

Return identityoperator(a.basis_l).