Full API

Autogenerated API list

CavityWaveguide.CavityWaveguideAbsorption — Type

CavityWaveguideAbsorption{B1,B2} <: CavityWaveguideOperator{B1,B2}

Structure for fast simultaneous cavity creation and waveguide annihilation operator

CavityWaveguide.CavityWaveguideEmission — Type

CavityWaveguideEmission{B1,B2} <: CavityWaveguideOperator{B1,B2}

Structure for fast simultaneous cavity annihilation and waveguide creation operator

CavityWaveguide.CavityWaveguideOperator — Type

Abstract type used for operators on acting on a combined WaveguideBasis and cavity basis (FockBasis)

CavityWaveguide.Detector — Type

Detector(wa,wb)

Detector operation defined by giving waveguide annihilation operator wa and wb from two subsystems. wa acts on the first subsystem of a LazyTensorKet or LazySumKet and wb on the second.

CavityWaveguide.InputOutputTimestepView — Type

InputOutputTimestepView{T} <:AbstractVector{T}

Structure for viewing slice along same times in one photon input output state.

CavityWaveguide.InputOutputView — Type

InputOutputView{T} <: AbstractMatrix{T}

Structure for viewing state with one photon in input and output waveguide.

CavityWaveguide.InputOutputWaveguideBasis — Type

InputOutputWaveguideBasis(N,times)

Basis for two waveguides containing in total N excitations. The two waveguides are refered to as input and output, but they are identical. Currently restricted to either N=1 or N=2. Times is timeinterval over which the photon state should be binned.

CavityWaveguide.LazySumKet — Type

LazySumKet(kets...)

Lazy sum of LazyTensorKets that is used to express entanglement between subsystems in LazyTensorKets.

CavityWaveguide.LazyTensorBra — Type

LazyTensorBra(bras)

Lazy tensor product between bras. Used in functions for beamsplitter and subsequent detection.

CavityWaveguide.LazyTensorKet — Type

LazyTensorKet(kets)

Lazy tensor product between kets. Used in functions for beamsplitter and subsequent detection.

CavityWaveguide.OnePhotonView — Method

OnePhotonView(ψ::Ket)
OnePhotonView(ψ::Ket,type)
OnePhotonView(ψ::Ket,index)
OnePhotonView(ψ::Ket,index,typw)

Return a view of the onephoton mode $ξ(t)$ given a state defined on a WaveguideBasis or InputOutputWaveguideBasis. If the state is a InputOutputWaveguideBasis the type parameter can be used to choose between the input or output mode with type = :input or type = :output If no index is provided the ground state is returned. Index should follow same form outlined in view_waveguide.

CavityWaveguide.TwoPhotonTimestepView — Type

TwoPhotonTimestepView{T}

Structure for viewing slice along same times in twophoton state.

CavityWaveguide.TwoPhotonView — Type

TwophotonView{T} <: AbstractMatrix{T}

Structure for viewing twophoton state as symmetric matrix (only upper triangluar part is stored in memory).

CavityWaveguide.WaveguideBasis — Type

WaveguideBasis(N, times)

Basis for time binned Waveguide where N is the number of photons in the waveguide. Currently restricted to either 1 or 2. Times is timeinterval over which the photon state should be binned.

CavityWaveguide.WaveguideCreate — Type

WaveguideCreate{N} <: WaveguideOperator

Operator structure for dispatching creation operation on Waveguide state. N is used to dispatch one or two photon routine.

CavityWaveguide.WaveguideDestroy — Type

WaveguideDestroy{N} <: WaveguideOperator

Operator structure for dispatching annihilation operation on Waveguide state. N is used to dispatch one or two photon routine.

CavityWaveguide.WaveguideOperator — Type

Abstract class for WaveguideOperators. Used to dispatch special mul! function.

CavityWaveguide.absorption — Method

absorption(b1::WaveguideBasis{T},b2::FockBasis) where T
absorption(b1::FockBasis,b2::WaveguideBasis{T}) where T

Create CavityWaveguideAbsorption that applies create(b::FockBasis) on FockBasis and destroy(b::WaveguideBasis{T}) on WaveguideBasis{T}.

CavityWaveguide.detect_double_click — Method

detect_double_click(ψ,detector1,detector2,projection)
detect_double_click(ψ,detector1,detector2)

Calculate probability of observing projection after beamsplitter operation and two subsequent detection events defined in detector1 and detector2 on the state ψ.

Arguments

ψ can be either LazyTensorKet or LazySumKet and is the state on which the beamsplitter and detection is applied

detector1 defines the first beamsplitter and subsequent detection operation. See Detector for more details on how to define.

detector2 defines the second beamsplitter and subsequent detection operation. See Detector for more details on how to define.

projection if given is a LazyTensorKet or LazySumKet which projects onto the state after the measurement. If no projection is given, instead the total probability of having the detector click is given by applying all possible combinations of projections using get_all_projectors.

Returns

If projection is given: Returns probability of having detector1 and detector2 click and being in state defined by projection

If projection is not given: Returns the total probability of having detector1 and detector2 click by applying all possibile projections with zerophotons in the waveguide using get_all_projectors.

See Beamsplitter for an example on how to use.

CavityWaveguide.detect_single_click — Method

detect_single_click(ψ,detector::Detector,projection)
detect_single_click(ψ,detector::Detector)

Calculate probability of observing projection after beamsplitter operation and subsequent detection event defined in detector on the state ψ.

Arguments

ψ can be either LazyTensorKet or LazySumKet and is the state on which the beamsplitter and detection is applied

detector defines the beamsplitter and subsequent detection operation. See Detector for more details on how to define.

projection if given is a LazyTensorKet or LazySumKet which projects onto the state after the measurement. If no projection is given, instead the total probability of having the detector click is given by applying all possible combinations of projections using get_all_projectors.

Returns

If projection is given returns probability of having detector click and being in state defined by projection

If projection is not given returns the total probability of having a the detector click (only a single click, for double clicks use detect_double_click) by applying all possibile projections with zerophotons in the waveguide using get_all_projectors.

See Beamsplitter for an example on how to use.

CavityWaveguide.emission — Method

emission(b1::WaveguideBasis{T},b2::FockBasis) where T
emission(b1::FockBasis,b2::WaveguideBasis{T}) where T

Create CavityWaveguideEmission that applies destroy(b::FockBasis) on FockBasis and create(b::WaveguideBasis{T}) on WaveguideBasis{T}.

CavityWaveguide.get_all_projectors — Method

get_all_projectors(b)

Returns all combinations of possible states with zerophotons in the waveguide.

CavityWaveguide.get_nsteps — Method

get_nsteps(basis::WaveguideBasis)
get_nsteps(basis::Basis)
get_nsteps(basis::CompositeBasis)

Return nsteps of WaveguideBasis given either a WaveguideBasis or a CompositeBasis containing a WaveguideBasis

CavityWaveguide.get_waveguide_basis — Method

get_waveguide_basis(basis::CompositeBasis)

Returns WaveguideBasis from CompositeBasis.bases

CavityWaveguide.get_waveguide_location — Method

get_waveguide_location(basis::WaveguideBasis)
get_waveguide_location(basis::CompositeBasis)

Return index of WaveguideBasis location in Hilbert space of basis b. Btotal = BW ⊗ BC where BW is a WaveguideBasis and BC some other basis then get_waveguide_location(Btotal) returns 1. While Btotal = BC ⊗ BW with get_waveguide_location(Btotal) returns 2.

CavityWaveguide.get_waveguide_operators — Method

get_waveguide_operators(basis::LazySum)
get_waveguide_operators(basis::LazyProduct)
get_waveguide_operators(basis::LazyTensor)
get_waveguide_operators(basis::Tuple)
get_waveguide_operators(basis::Array)
get_waveguide_operators(basis::WaveguideOperator)

Returns all WaveguideOperator in LazyOperator or from a list of operators. If no WaveguideOperator is found, and empty array is returned.

CavityWaveguide.get_waveguidetimeindex — Method

get_waveguidetimeindex(op)

Return timeindex of operator or list of operators containing WaveguideOperator and assert that all timeindeces are the same.

CavityWaveguide.get_waveguidetimeindex — Method

get_waveguidetimeindex(basis::WaveguideBasis)
get_waveguidetimeindex(basis::Basis)
get_waveguidetimeindex(basis::CompositeBasis)

Return timeindex of WaveguideBasis given either a WaveguideBasis or a CompositeBasis containing a WaveguideBasis

CavityWaveguide.inputabsorption — Method

inputabsorption(b1::InputOutputWaveguideBasis{T},b2::FockBasis) where T
inputabsorption(b1::FockBasis,b2::InputOutputWaveguideBasis{T}) where T

Create CavityWaveguideAbsorption that applies create(b::FockBasis) on FockBasis and inputdestroy(b::InputOutputWaveguideBasis{T}) on InputOutputWaveguideBasis{T}.

CavityWaveguide.inputcreate — Method

inputcreate(basis::WaveguideBasis{N}) where N

Creation operator for input waveguide in InputOutputWaveguideBasis for either one or two photons.

CavityWaveguide.inputdestroy — Method

inputdestroy(basis::WaveguideBasis{N})

Annihilation operator for input waveguide in InputOutputWaveguideBasis for either one or two photons.

CavityWaveguide.inputemission — Method

inputemission(b1::InputOutputWaveguideBasis{T},b2::FockBasis) where T
inputemission(b1::FockBasis,b2::InputOutputWaveguideBasis{T}) where T

Create CavityWaveguideEmission that applies destroy(b::FockBasis) on FockBasis and inputcreate(b::InputOutputWaveguideBasis{T}) on InputOutputWaveguideBasis{T}.

CavityWaveguide.onephoton — Method

onephoton(b::InputOutputWaveguideBasis,type,ξ::Function,times,args...,norm=True)
onephoton(b::InputOutputWaveguideBasis,type,ξ;norm=true)

Create a onephoton wavepacket of the form $W_{i/o}^\dagger(\xi) |\emptyset,\emptyset \rangle = \int_{t_0}^{t_{end}} dt \xi(t) w_{i/o}^\dagger(t) |\emptyset,\emptyset \rangle$ where the subscript i/o denotes the input or output waveguide and determined by the type parameter.

Input

b is the basis of the input output waveguides.

type determines which waveguide the onephoton state is created in. Can be either :input or :outputput.

ξ can either be a function evaluated as ξ.(times,args...) or a vector of length: b.nsteps.

If norm==true the state is normalized through normalize!.

CavityWaveguide.onephoton — Method

onephoton(b::WaveguideBasis,ξ::Function,times,args...,norm=True)
onephoton(b::WaveguideBasis,ξvec;norm=true)

Create a onephoton wavepacket of the form $W^\dagger(\xi) |0⟩ = \int_{t_0}^{t_{end}} dt \xi(t) w^\dagger(t) |0⟩$. Here $t_0=0$ and $t_{end}$ is determined by WaveguideBasis. ξ is a function evaluated as ξ.(times,args...). ξvec is a vector of length: b.nsteps. If norm==true the state is normalized through normalize!.

CavityWaveguide.outputabsorption — Method

outputabsorption(b1::InputOutputWaveguideBasis{T},b2::FockBasis) where T
outputabsorption(b1::FockBasis,b2::InputOutputWaveguideBasis{T}) where T

Create CavityWaveguideAbsorption that applies create(b::FockBasis) on FockBasis and outputdestroy(b::InputOutputWaveguideBasis{T}) on InputOutputWaveguideBasis{T}.

CavityWaveguide.outputcreate — Method

Outputcreate(basis::WaveguideBasis{N}) where N

Creation operator for output waveguide in InputOutputWaveguideBasis for either one or two photons.

CavityWaveguide.outputdestroy — Method

outputdestroy(basis::WaveguideBasis{N})

Annihilation operator for output waveguide in InputOutputWaveguideBasis for either one or two photons.

CavityWaveguide.outputemission — Method

outputemission(b1::InputOutputWaveguideBasis{T},b2::FockBasis) where T
outputemission(b1::FockBasis,b2::InputOutputWaveguideBasis{T}) where T

Create CavityWaveguideEmission that applies destroy(b::FockBasis) on FockBasis and outputcreate(b::InputOutputWaveguideBasis{T}) on InputOutputWaveguideBasis{T}.

CavityWaveguide.plot_twophoton! — Method

plot_twophoton!(ax,twophotonstate::TwophotonView,times)
plot_twophoton!(ax,state::Ket,times)

Plots the twophoton state in the given ax. If state is a Ket TwoPhotonView is called to extract twophotonstate. Returns ax.contour object.

CavityWaveguide.set_waveguidetimeindex! — Method

set_waveguidetimeindex!(op,index)

Set timeindex of all WaveguideOperator in operator or list of operators to index

CavityWaveguide.twophoton — Method

twophoton(b::InputOutputWaveguideBasis,type,ξ::Function,times,args...,norm=True)
twophoton(b::InputOutputWaveguideBasis,type,ξ;norm=true)

Create a twophoton wavepacket of the form $\int_{t_0}^{t_{end}} dt' \int_{t_0}^{t_{end}} dt \xi(t,t') w_{i/o}^\dagger(t)w_{i/o}^\dagger(t') |\emptyset,\emptyset \rangle$ The subscript i/o denotes the input or output waveguide and is determined by the type parameter.

Input

b is the basis of the input output waveguides.

type determines which waveguide the twophoton state is created in. Can be either :input or :outputput or :inputoutput for an entangled onephoton state in the input waveguide and onephoton state in the outputwaveguide.

ξ can either be a function evaluated as ξ.(times,args...) or a vector of length: b.nsteps.

If norm==true the state is normalized through normalize!.

CavityWaveguide.twophoton — Method

twophoton(b::WaveguideBasis,ξ::Function,times,args...,norm=True)
twophoton(b::WaveguideBasis,ξvec::Matrix;norm=true)

Create a twophoton wavepacket of the form $\int_{t_0}^{t_{end}} dt' \int_{t_0}^{t_{end}} dt \xi(t,t') w^\dagger(t)w^\dagger(t') |0⟩$. Here $t_0=0$ and $t_{end}$ is determined by WaveguideBasis. ξ is a function evaluated as ξ(t1,t2,args...). ξvec is a matrix of dimension: (b.nsteps,b.nsteps), where ξvec[i,j] = ξ(times[i],times[j]), where times is defined in WaveguideBasis.

CavityWaveguide.view_waveguide — Method

view_waveguide(ψ::ket)
view_waveguide(ψ::ket,index)

View the Waveguide state given a state ψ containing a WaveguideBasis by returning view(reshape(ψ.data,Tuple(ψ.basis.shape)),index...). If no index is provided the ground state is returned. The index provided should be of the form [:,i,j] where (:) is at the location of the WaveguideBasis and i and j are indeces of other basises. See example:

times=0:0.1:10
bw = WaveguideBasis(2,times)
bc1 = FockBasis(2)
bc2 = FockBasis(2)
ψ_waveguide = onephoton(bw,x->1)
ψ_total = ψ_waveguide ⊗ fockstate(bc1,1) ⊗ fockstate(bc2,1)
ψ_view = view_waveguide(ψ_total)
ψ_view_index = view_waveguide(ψ_total,[:,1,1])
ψ_view==ψ_view_index

CavityWaveguide.waveguide_evolution — Method

waveguide_evolution(tspan, psi0, H; fout)

Integrate time-dependent Schroedinger equation to evolve states or compute propagators.

Arguments

tspan: Vector specifying the points of time for which output should be displayed.

psi0: Initial state vector can only be a ket.

H: Operator containing a WaveguideOperator either through a LazySum or LazyTensor.

fout=nothing: If given, this function fout(t, psi) is called every time step. Example: fout(t,psi) = expect(A,psi) will return the epectation value of A at everytimestep. ATTENTION: The state psi is neither normalized nor permanent! It is still in use by the ode solver and therefore must not be changed.

Output

if fout=nothing the output of the solver will be the state ψ at the last timestep.

if fout is given a tuple with the state ψ at the last timestep and the output of fout is given. If fout returns a tuple the tuple will be flattened.

Example fout(t,psi) = (expect(A,psi),expect(B,psi)) will result in a tuple (ψ, ⟨A(t)⟩,⟨B(t)⟩), where ⟨A(t)⟩ is a vector with the expectation value of A as a function of time.

CavityWaveguide.waveguide_montecarlo — Method

waveguide_montecarlo(times,psi,H,J;fout=nothing)

CavityWaveguide.zerophoton — Method

zerophoton(bw::InputOutputWaveguideBasis)

Create a waveguide vacuum state $|\emptyset,\emptyset \rangle$ of the two waveguides.

CavityWaveguide.zerophoton — Method

zerophoton(bw::WaveguideBasis)

Create a waveguide vacuum state |0⟩

QuantumInterface.dagger — Method

dagger(op::WaveguideCreate)
dagger(op::WaveguideDestroy)

Dagger opration on Waveguide operator.

QuantumInterface.dagger — Method

dagger(op::WaveguideCreate)
dagger(op::WaveguideCreate)

Dagger opration on Waveguide operator.

QuantumInterface.tensor — Method

tensor(op1::AbstractOperator,op2::CavityWaveguideAbsorption)
tensor(op1::CavityWaveguideAbsorption,op2::AbstractOperator)
tensor(op1::AbstractOperator,op2::CavityWaveguideEmission)
tensor(op1::CavityWaveguideEmission,op2::AbstractOperator)

Methods for tensorproducts between QuantumOptics.jl operators and CavityWaveguideOperator.

QuantumOpticsBase.create — Method

create(basis::WaveguideBasis{1})
create(basis::WaveguideBasis{2})

Creation operator for WaveguideBasis for either one or two photons.

QuantumOpticsBase.destroy — Method

destroy(basis::WaveguideBasis{1})
destroy(basis::WaveguideBasis{2})

Annihilation operator for WaveguideBasis for either one or two photons.

QuantumOpticsBase.identityoperator — Method

identityoperator(a::CavityWaveguideOperator)

Return identityoperator(a.basisl). QUESTION: (does basisl or basis_r matter?)

QuantumOpticsBase.identityoperator — Method

identityoperator(a::WaveguideOperator)

Return identityoperator(a.basis_l).

LinearAlgebra.mul! — Function

mul!(result::Ket{B1}, a::LazyTensor{B1,B2,F,I,T}, b::Ket{B2}, alpha, beta)
mul!(result::Bra{B1}, a::Bra{B2}, b::LazyTensor{B1,B2,F,I,T}, alpha, beta)

In-place multiplication of operators/state vectors. Updates result as result = alpha*a*b + beta*result. a is a LazyTensor that contains a WaveguideOperator

mul!(result::Ket{B1}, a::CavityWaveguideOperator, b::Ket{B2}, alpha, beta) where {B1<:Basis,B2<:Basis}
mul!(result::Bra{B1}, a::Bra{B2}, b::CavityWaveguideOperator, alpha, beta) where {B1<:Basis,B2<:Basis}

Fast in-place multiplication of operators/state vectors. Updates result as result = alpha*a*b + beta*result. a is a CavityWaveguideOperator. Routine only works if [WaveguideBasis] is the first or last basis in Hilbertspace.

QuantumOpticsBase.create — Function

create(basis::WaveguideBasis{1})
create(basis::WaveguideBasis{2})

Creation operator for WaveguideBasis for either one or two photons.

QuantumOpticsBase.destroy — Function

destroy(basis::WaveguideBasis{1})
destroy(basis::WaveguideBasis{2})

Annihilation operator for WaveguideBasis for either one or two photons.

QuantumInterface.dagger — Function

dagger(op::WaveguideCreate)
dagger(op::WaveguideCreate)

Dagger opration on Waveguide operator.

dagger(op::WaveguideCreate)
dagger(op::WaveguideDestroy)

Dagger opration on Waveguide operator.

QuantumInterface.tensor — Function

tensor(op1::AbstractOperator,op2::CavityWaveguideAbsorption)
tensor(op1::CavityWaveguideAbsorption,op2::AbstractOperator)
tensor(op1::AbstractOperator,op2::CavityWaveguideEmission)
tensor(op1::CavityWaveguideEmission,op2::AbstractOperator)

Methods for tensorproducts between QuantumOptics.jl operators and CavityWaveguideOperator.

QuantumOpticsBase.identityoperator — Function

identityoperator(a::WaveguideOperator)

Return identityoperator(a.basis_l).

identityoperator(a::CavityWaveguideOperator)

Return identityoperator(a.basisl). QUESTION: (does basisl or basis_r matter?)